ai/requirements.txt
fastapi
uvicorn
langchain
langchain-community
transformers
sentence-transformers
faiss-cpu
youtube-transcript-api
python-dotenv